requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
assemblyai==0.36.0
attrs==24.3.0
av==14.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.8
ctranslate2==4.5.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.9.0.post1
faster-whisper==1.1.1
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.8
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.24.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
onnxruntime==1.20.1
openai==1.59.5
orjson==3.10.13
packaging==24.2
pillow==11.1.0
playsound==1.2.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-dotenv==1.0.1
pyttsx3==2.98
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
setuptools==75.6.0
# simpleaudio==1.0.4
smallestai==1.3.4
sniffio==1.3.1
SQLAlchemy==2.0.36
srt==3.5.3
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
twilio==9.4.2
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.1
Werkzeug==3.1.3
yarl==1.18.3
gunicorn==20.1.0


huggingface-hub>=0.23.0

flask
streamlit
ctransformers
sentence-transformers
pinecone-client
langchain
langchain_community
langchain_huggingface
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.31.1
google-cloud-texttospeech==2.25.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
pydub==0.25.1

pypdf


-e .